Refactor EditSignup code
