--- packages/ilmomasiina-frontend/src/routes/EditSignup/EditSignup.scss
@@ -1,8 +1,0 @@
-.EditSignup--wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  max-width: 600px;
-  margin: 4rem auto;
-}
