+++ packages/ilmomasiina-frontend/src/routes/EditSignup/components/DeleteSignup/index.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
 
-import { Col, Row } from 'react-bootstrap';
 import { shallowEqual } from 'react-redux';
 
 import { deleteSignup } from '../../../../modules/editSignup/actions';
 import { useTypedDispatch, useTypedSelector } from '../../../../store/reducers';
+import NarrowContainer from '../NarrowContainer';
 
 const DELETE_CONFIRM_MS = 4000;
 
@@ -36,29 +36,27 @@ const DeleteSignup = ({ editToken }: Props) => {
   }
 
   return (
-    <Row className="justify-content-md-center text-center my-5">
-      <Col xs="12" md="10" lg="8">
-        <h2>Poista ilmoittautuminen</h2>
-        <p>
-          Oletko varma, että haluat poistaa ilmoittautumisesi tapahtumaan
-          {' '}
-          <strong>
-            {event!.title}
-          </strong>
-          ?
-        </p>
-        <p>
-          Jos poistat ilmoittautumisesi, menetät paikkasi jonossa. Jos
-          muutat mielesi, voit aina ilmoittautua tapahtumaan uudelleen
-          myöhemmin, mutta siirryt silloin jonon hännille.
-          {' '}
-          <strong>Tätä toimintoa ei voi perua.</strong>
-        </p>
-        <button type="button" disabled={submitting} onClick={onDeleteClick} className="btn btn-danger">
-          {confirming ? 'Paina uudelleen varmistukseksi\u2026' : 'Poista ilmoittautuminen'}
-        </button>
-      </Col>
-    </Row>
+    <NarrowContainer className="text-center my-5">
+      <h2>Poista ilmoittautuminen</h2>
+      <p>
+        Oletko varma, että haluat poistaa ilmoittautumisesi tapahtumaan
+        {' '}
+        <strong>
+          {event!.title}
+        </strong>
+        ?
+      </p>
+      <p>
+        Jos poistat ilmoittautumisesi, menetät paikkasi jonossa. Jos
+        muutat mielesi, voit aina ilmoittautua tapahtumaan uudelleen
+        myöhemmin, mutta siirryt silloin jonon hännille.
+        {' '}
+        <strong>Tätä toimintoa ei voi perua.</strong>
+      </p>
+      <button type="button" disabled={submitting} onClick={onDeleteClick} className="btn btn-danger">
+        {confirming ? 'Paina uudelleen varmistukseksi\u2026' : 'Poista ilmoittautuminen'}
+      </button>
+    </NarrowContainer>
   );
 };
